+++ NEWS.md
@@ -1,5 +1,6 @@
 # BayesMallows (development versions)
 
+* Fixed bug in assign_cluster() with burn-in and thinning. Issue #433 on GitHub.
 * heat_plot() function now works with clusters.
 * An error in plot(mod, parameter = "rho") when run with parallel chains has 
   been fixed. It previously showed unnormalized probabilities, but now shows